backend/requirements.txt
# backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas
numpy
sympy
scipy
yfinance
pymongo
dnspython